Adding riflecore dispersion values
